TODO: Pembelian Invoice

+++ app/Services/Core/Purchase/PurchaseOrderService.php
@@ -1,22 +1,30 @@
 <?php
+
 namespace App\Services\Core\Purchase;
 
+use App\Exceptions\PurchaseOrderException;
 use App\Models\Pembelian;
 use App\Models\PembelianDetail;
+use App\Models\ProdukPersediaan;
+use App\Models\ProdukPersediaanDetail;
+use App\Models\ProdukVarianHarga;
 use App\Models\Transaksi;
+use App\Services\Core\Jurnal\JurnalService;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Validator;
 
 class PurchaseOrderService
 {
-    public function storePurchaseOrder($request) {
+    use JurnalService;
+    public function storePurchaseOrder($request)
+    {
         $rules = [
             'id_kontak' => 'required|numeric',
             'transaksi_no' => 'nullable|string',
             'tanggal' => 'required|date_format:Y-m-d H:i:s',
             'id_gudang' => 'required|numeric',
             'diskon' => 'nullable|numeric',
-            'catatan' => 'nullable|numeric',
+            'catatan' => 'nullable|string',
             'pembelianDetail' => 'required|array',
             'pembelianDetail.*.kode_produkvarian' => 'required|string',
             'pembelianDetail.*.qty' => 'required|numeric',
@@ -51,13 +59,13 @@ class PurchaseOrderService
                     'qty' => $item['qty'],
                     'diskonjenis' => 'persen',
                     'diskon' => $item['diskon'] ?: 0,
-                    'total' => (int)($item['harga']*$item['qty']*(1-($item['diskon'] ?: 0)/100)),
+                    'total' => (int)($item['harga'] * $item['qty'] * (1 - ($item['diskon'] ?: 0) / 100)),
                     'totalraw' => (int)$item['harga'] * $item['qty'],
                     'id_gudang' => $request['id_gudang']
                 ]);
-                $rawTotal += (int)($item['harga']*$item['qty']*(1-($item['diskon'] ?: 0)/100));
+                $rawTotal += (int)($item['harga'] * $item['qty'] * (1 - ($item['diskon'] ?: 0) / 100));
             }
-            $grandTotal = (int)($rawTotal * (1-$request['diskon']/100));
+            $grandTotal = (int)($rawTotal * (1 - $request['diskon'] / 100));
             $pembelian->update([
                 'totalraw' => $rawTotal,
                 'grandtotal' => $grandTotal,
@@ -71,7 +79,8 @@ class PurchaseOrderService
             throw $e;
         }
     }
-    public function updatePurchaseOrder($idPembelian, $request) {
+    public function updatePurchaseOrder($idPembelian, $request)
+    {
         $rules = [
             'id_kontak' => 'required|numeric',
             'tanggal' => 'required|date_format:Y-m-d H:i:s',
@@ -125,8 +134,8 @@ class PurchaseOrderService
                             $newData['id_gudang'] = $request['id_gudang'];
                         }
                         if (!empty($newData)) {
-                            $newData['totalraw'] = $newData['qty']*$newData['harga'];
-                            $newData['total'] = $newData['qty']*$newData['harga']*(1-$newData['diskon']/100);
+                            $newData['totalraw'] = $newData['qty'] * $newData['harga'];
+                            $newData['total'] = $newData['qty'] * $newData['harga'] * (1 - $newData['diskon'] / 100);
                             PembelianDetail::where('id_pembeliandetail', $key)->update($newData);
                         }
                     } else {
@@ -137,12 +146,12 @@ class PurchaseOrderService
                             'qty' => $item['qty'],
                             'diskonjenis' => 'persen',
                             'diskon' => $item['diskon'] ?: 0,
-                            'total' => (int)($item['harga']*$item['qty']*(1-($item['diskon'] ?: 0)/100)),
+                            'total' => (int)($item['harga'] * $item['qty'] * (1 - ($item['diskon'] ?: 0) / 100)),
                             'totalraw' => (int)$item['harga'] * $item['qty'],
                             'id_gudang' => $request['id_gudang']
                         ]);
                     }
-                } 
+                }
             }
             $pembelian->refresh();
             $rawTotal = 0;
@@ -151,7 +160,7 @@ class PurchaseOrderService
             }
             $pembelian->update([
                 'totalraw' => $rawTotal,
-                'grandtotal' => $rawTotal*(1-$pembelian['diskon']/100)
+                'grandtotal' => $rawTotal * (1 - $pembelian['diskon'] / 100)
             ]);
             $pembelian->refresh();
             DB::commit();
@@ -161,12 +170,16 @@ class PurchaseOrderService
             throw $e;
         }
     }
-    public function deletePurchaseOrder($idPembelian) {
+    public function deletePurchaseOrder($idPembelian)
+    {
         if ((new Pembelian())->where('id_pembelian', $idPembelian)->has('pembelianInvoice')->first()) {
             abort(403);
         }
         DB::beginTransaction();
         try {
+            if (!Pembelian::where('id_pembelian', $idPembelian)->has('pembelianInvoice')->first()) {
+                throw new PurchaseOrderException('Pembelian order sudah di-invoice, penghapusan tidak diizinkan');
+            }
             PembelianDetail::where('id_pembelian', $idPembelian)->delete();
             $pembelian = Pembelian::where('id_pembelian', $idPembelian)->first();
             $pembelian->delete();
@@ -177,10 +190,107 @@ class PurchaseOrderService
             throw $e;
         }
     }
-    public function toInvoicePurchaseOrder($idPembelian) {
-        $pembelian = (new Pembelian())->with(['kontak','pembelianDetail' => function ($rel) {
-            $rel->leftJoin(DB::raw("(select id_pembeliandetailparent, sum(qty) as jumlah_diinvoice from toko_griyanaura.tr_pembeliandetail where id_pembeliandetailparent is not null group by id_pembeliandetailparent) as x"), 'x.id_pembeliandetailparent', 'toko_griyanaura.tr_pembeliandetail.id_pembeliandetail');
-        } ])->where('id_pembelian', $idPembelian)->first();
-        
+    public function storeToInvoicePurchaseOrder($idPembelian, $request)
+    {
+        DB::beginTransaction();
+        try {
+            $pembelian = (new Pembelian())->with(['kontak', 'pembelianDetail' => function ($rel) {
+                $rel->leftJoin(DB::raw("(select id_pembeliandetailparent, sum(qty) as jumlah_diinvoice from toko_griyanaura.tr_pembeliandetail where id_pembeliandetailparent is not null group by id_pembeliandetailparent) as x"), 'x.id_pembeliandetailparent', 'toko_griyanaura.tr_pembeliandetail.id_pembeliandetail');
+            }])->where('id_pembelian', $idPembelian)->first();
+            $noTransaksi = DB::select("select ('PI' || lpad(nextval('toko_griyanaura.tr_pembelian_invoice_seq')::varchar, 6, '0')) as no_transaksi")[0]->no_transaksi;
+            $transaksi = Transaksi::create([
+                'transaksi_no' => $noTransaksi,
+                'id_transaksijenis' => 'pembelian_invoice',
+            ]);
+            $pembelianInvoice = Pembelian::create([
+                'id_transaksi' => $transaksi->id_transaksi,
+                'id_kontak' => $pembelian->id_kontak,
+                'jenis' => 'invoice',
+                'transaksi_no' => $noTransaksi,
+                'tanggal' => $pembelian->tanggal,
+                'tanggaltempo' => $request['tanggaltempo'],
+                'catatan' => $pembelian->catatan,
+                'diskonjenis' => 'persen',
+                'diskon' => $pembelian->diskon ?: 0,
+                'pembelian_parent' => $pembelian->id_pembelian,
+                'id_gudang' => $pembelian->id_gudang
+            ]);
+            $pembelianDetail = $pembelian->pembelianDetail->keyBy('id_pembeliandetail');
+            $rawTotal = 0;
+            foreach ($request['pembelianDetail'] as $key => $item) {
+                if (isset($item['check'])) {
+                    if (isset($pembelianDetail[$key]) and $pembelianDetail[$key]['qty'] - $pembelianDetail[$key]['jumlah_diinvoice'] >= $item['qty']) {
+                        PembelianDetail::create([
+                            'id_pembelian' => $pembelianInvoice->id_pembelian,
+                            'kode_produkvarian' => $pembelianDetail[$key]['kode_produkvarian'],
+                            'harga' => (int)$pembelianDetail[$key]['harga'],
+                            'qty' => $item['qty'],
+                            'diskonjenis' => 'persen',
+                            'diskon' => $pembelianDetail[$key]['diskon'] ?: 0,
+                            'total' => (int)($pembelianDetail[$key]['harga'] * $item['qty'] * (1 - ($pembelianDetail[$key]['diskon'] ?: 0) / 100)),
+                            'totalraw' => (int)$pembelianDetail[$key]['harga'] * $item['qty'],
+                            'id_gudang' => $pembelian->id_gudang,
+                            'id_pembeliandetailparent' => $key
+                        ]);
+                        $rawTotal += (int)($pembelianDetail[$key]['harga'] * $item['qty'] * (1 - ($pembelianDetail[$key]['diskon'] ?: 0) / 100));
+                    }
+                }
+            }
+            $grandTotal = (int)($rawTotal * (1 - $pembelian['diskon'] / 100));
+            $pembelianInvoice->update([
+                'totalraw' => $rawTotal,
+                'grandtotal' => $grandTotal
+            ]);
+            $pembelianInvoice = Pembelian::with('pembelianDetail.produkVarian.produk')->find($pembelianInvoice->id_pembelian);
+            $detailTransaksi = [];
+            foreach ($pembelianInvoice->pembelianDetail as $item) {
+                if ($item->produk->in_stok == true) {
+                    $persediaanProduk = ProdukPersediaan::where('kode_produkvarian', $item['kode_produkvarian'])->where('id_gudang', $item['id_gudang'])->first();
+                    if (!$persediaanProduk) {
+                        if (ProdukVarianHarga::where('kode_produkvarian', $item['kode_produkvarian'])->join('toko_griyanaura.ms_produkharga as ph', 'ph.id_produkharga', 'toko_griyanaura.ms_produkvarianharga.id_produkharga')->where('ph.id_varianharga', $item['id_gudang'])->first()) {
+                            $defaultVarianHarga = ProdukVarianHarga::where('kode_produkvarian', $item['kode_produkvarian'])->join('toko_griyanaura.ms_produkharga as ph', 'ph.id_produkharga', 'toko_griyanaura.ms_produkvarianharga.id_produkharga')->where('ph.id_varianharga', $item['id_gudang'])->first()->id_produkvarianharga;
+                        } else {
+                            $defaultVarianHarga = ProdukVarianHarga::where('kode_produkvarian', $item['kode_produkvarian'])->join('toko_griyanaura.ms_produkharga as ph', 'ph.id_produkharga', 'toko_griyanaura.ms_produkvarianharga.id_produkharga')->where('ph.id_varianharga', 1 /* Reguler */)->first()->id_produkvarianharga;
+                        }
+                        $persediaanProduk = ProdukPersediaan::create([
+                            'id_gudang' => $item['id_gudang'],
+                            'kode_produkvarian' => $item['kode_produkvarian'],
+                            'stok' => 0,
+                            'default_varianharga' => $defaultVarianHarga
+                        ]);
+                    }
+                    ProdukPersediaan::where('kode_produkvarian', $item->kode_produkvarian)->where('id_gudang', $item->id_gudang)->update([
+                        'stok' => DB::raw('stok +' . (int)$item->qty)
+                    ]);
+                    $dataPersediaanDetail = ProdukPersediaanDetail::create([
+                        'id_persediaan' => $persediaanProduk->id_persediaan,
+                        'tanggal' => $pembelianInvoice->tanggal,
+                        'keterangan' => "#{$pembelianInvoice->transaksi_no} Invoice Pembelian",
+                        'stok_in' => $item->qty,
+                        'hargabeli' => (int)($item->harga * (1-$item->diskon/100) * (1-$pembelianInvoice->diskon/100))
+                    ]);
+                }
+                $detailTransaksi[] = [
+                    'kode_akun' => $item->produkVarian->produk->default_akunpersediaan,
+                    'keterangan' => $item->produkVarian->varian,
+                    'nominaldebit' => (int)($item->qty * $item->harga * (1-$item->diskon/100) * (1-$pembelianInvoice->diskon/100)),
+                    'nominalkredit' => 0
+                ];
+            }
+            $detailTransaksi[] = [
+                'kode_akun' => '2001',
+                'keterangan' => null,
+                'nominaldebit' => 0,
+                'nominalkredit' => $pembelianInvoice->grandtotal
+            ];
+            if (isset($detailTransaksi)) {
+                $this->entryJurnal($pembelianInvoice->id_transaksi, $detailTransaksi);
+            }
+            DB::commit();
+            return $pembelianInvoice;
+        } catch (\Exception $e) {
+            throw $e;
+            DB::rollBack();
+        }
     }
 }
